+++ netbox/netbox/choices.py
@@ -7,8 +7,10 @@ __all__ = (
     'ButtonColorChoices',
     'ColorChoices',
     'CSVDelimiterChoices',
+    'DistanceUnitChoices',
     'ImportFormatChoices',
     'ImportMethodChoices',
+    'WeightUnitChoices',
 )
 
 
@@ -157,3 +159,39 @@ class CSVDelimiterChoices(ChoiceSet):
         (SEMICOLON, _('Semicolon')),
         (TAB, _('Tab')),
     ]
+
+
+class DistanceUnitChoices(ChoiceSet):
+
+    # Metric
+    UNIT_KILOMETER = 'km'
+    UNIT_METER = 'm'
+
+    # Imperial
+    UNIT_MILE = 'mi'
+    UNIT_FOOT = 'ft'
+
+    CHOICES = (
+        (UNIT_KILOMETER, _('Kilometers')),
+        (UNIT_METER, _('Meters')),
+        (UNIT_MILE, _('Miles')),
+        (UNIT_FOOT, _('Feet')),
+    )
+
+
+class WeightUnitChoices(ChoiceSet):
+
+    # Metric
+    UNIT_KILOGRAM = 'kg'
+    UNIT_GRAM = 'g'
+
+    # Imperial
+    UNIT_POUND = 'lb'
+    UNIT_OUNCE = 'oz'
+
+    CHOICES = (
+        (UNIT_KILOGRAM, _('Kilograms')),
+        (UNIT_GRAM, _('Grams')),
+        (UNIT_POUND, _('Pounds')),
+        (UNIT_OUNCE, _('Ounces')),
+    )
